WIP integration test with a terraform provider binary.
TODO: this test requires the aws provider binary to present at a hardcoded
path inside the testdata directory. i need to make running this test configurable
since we may not want to always assume this dependency is satisfied, and it would
probably be better to use an env var to locate the plugin binary since i don't
plan to check it in to testdata.