enable provider repos to run the generator w/o a source checkout
in order to do this a new command was added to compile templates
from the hack/template directory into the packages in
internal/template/compiled and the filesystem-based
TemplateGetter implementation was replaced in main.go with one
which reads template data from these compiled modules. This means
that the compilation process must be run in order to update templates.
A hybrid TemplateGetter could be written which attempts to read from
the local filesystem, falling back to the compiled getter when run
outside of tree (ie hack directory is not present).

Signed-off-by: Kasey Kirkham <kasey@upbound.io>